# Conflicts:
#	dsa2000_cal/src/dsa2000_cal/iterative_calibrator.py
#	dsa2000_cal/src/dsa2000_fm/actors/aggregator.py
#	dsa2000_cal/src/dsa2000_fm/actors/calibration_solution_cache.py
#	dsa2000_cal/src/dsa2000_fm/actors/calibrator.py
#	dsa2000_cal/src/dsa2000_fm/actors/model_predictor.py
#	dsa2000_cal/src/dsa2000_fm/actors/system_gain_simulator.py
#	workflows/streaming_fm_distributed/ray/src/fm_main.py

+++ dsa2000_cal/src/dsa2000_common/common/mixed_precision_utils.py
@@ -90,7 +90,7 @@ X = TypeVar("X")
 class Policy:
     """Encapsulates casting for inputs, outputs and parameters."""
     vis_dtype: jnp.dtype = np.complex64
-    weight_dtype: jnp.dtype = np.float32
+    weight_dtype: jnp.dtype = np.float16
     flag_dtype: jnp.dtype = np.bool_
     image_dtype: jnp.dtype = np.float32
     gain_dtype: jnp.dtype = np.complex64
